Moved LambdaLever into its own files
