+++ wrapper/Convert/SireOpenMM/lambdalever.cpp
@@ -1,0 +1,90 @@
+/********************************************\
+  *
+  *  Sire - Molecular Simulation Framework
+  *
+  *  Copyright (C) 2023  Christopher Woods
+  *
+  *  This program is free software; you can redistribute it and/or modify
+  *  it under the terms of the GNU General Public License as published by
+  *  the Free Software Foundation; either version 3 of the License, or
+  *  (at your option) any later version.
+  *
+  *  This program is distributed in the hope that it will be useful,
+  *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  *  GNU General Public License for more details.
+  *
+  *  You should have received a copy of the GNU General Public License
+  *  along with this program; if not, write to the Free Software
+  *  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+  *
+  *  For full details of the license please see the COPYING file
+  *  that should have come with this distribution.
+  *
+  *  You can contact the authors via the website
+  *  at http://sire.openbiosim.org
+  *
+\*********************************************/
+
+#include "lambdalever.h"
+
+using namespace SireOpenMM;
+
+LambdaLever::LambdaLever() : SireBase::ConcreteProperty<LambdaLever, SireBase::Property>()
+{
+}
+
+LambdaLever::LambdaLever(const LambdaLever &other)
+    : SireBase::ConcreteProperty<LambdaLever, SireBase::Property>(other)
+{
+}
+
+LambdaLever::~LambdaLever()
+{
+}
+
+LambdaLever &LambdaLever::operator=(const LambdaLever &other)
+{
+    Property::operator=(other);
+    return *this;
+}
+
+bool LambdaLever::operator==(const LambdaLever &other) const
+{
+    return true;
+}
+
+bool LambdaLever::operator!=(const LambdaLever &other) const
+{
+    return not this->operator==(other);
+}
+
+LambdaLever *LambdaLever::clone() const
+{
+    return new LambdaLever(*this);
+}
+
+const char *LambdaLever::what() const
+{
+    return LambdaLever::typeName();
+}
+
+const char *LambdaLever::typeName()
+{
+    return QMetaType::typeName(qMetaTypeId<LambdaLever>());
+}
+
+void LambdaLever::set_lambda(OpenMM::Context &context,
+                             double lambda_value) const
+{
+    qDebug() << "set lambda to" << lambda_value;
+}
+
+void LambdaLever::set_force_index(const QString &force, int index)
+{
+}
+
+void LambdaLever::add_perturbable_molecule(const OpenMMMolecule &molecule,
+                                           const QHash<QString, qint32> &start_indicies)
+{
+}
