+++ wrapper/Convert/SireOpenMM/lambdalever.h
@@ -1,0 +1,84 @@
+/********************************************\
+  *
+  *  Sire - Molecular Simulation Framework
+  *
+  *  Copyright (C) 2023  Christopher Woods
+  *
+  *  This program is free software; you can redistribute it and/or modify
+  *  it under the terms of the GNU General Public License as published by
+  *  the Free Software Foundation; either version 3 of the License, or
+  *  (at your option) any later version.
+  *
+  *  This program is distributed in the hope that it will be useful,
+  *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  *  GNU General Public License for more details.
+  *
+  *  You should have received a copy of the GNU General Public License
+  *  along with this program; if not, write to the Free Software
+  *  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+  *
+  *  For full details of the license please see the COPYING file
+  *  that should have come with this distribution.
+  *
+  *  You can contact the authors via the website
+  *  at http://sire.openbiosim.org
+  *
+\*********************************************/
+
+#ifndef SIREOPENMM_LAMBDALEVER_H
+#define SIREOPENMM_LAMBDALEVER_H
+
+#include "openmmmolecule.h"
+
+#include "SireCAS/expression.h"
+#include "SireCAS/symbol.h"
+
+SIRE_BEGIN_HEADER
+
+namespace SireOpenMM
+{
+
+    /** This is a lever that is used to change the parameters in an OpenMM
+     *  context according to a lambda value. This is actually a collection
+     *  of levers, each of which is controlled by the main lever.
+     *
+     *  You can use SireCAS expressions to control how each lever changes
+     *  each parameter
+     */
+    class LambdaLever : public SireBase::ConcreteProperty<LambdaLever, SireBase::Property>
+    {
+    public:
+        LambdaLever();
+        LambdaLever(const LambdaLever &other);
+
+        virtual ~LambdaLever();
+
+        LambdaLever &operator=(const LambdaLever &other);
+
+        bool operator==(const LambdaLever &other) const;
+        bool operator!=(const LambdaLever &other) const;
+
+        LambdaLever *clone() const;
+
+        const char *what() const;
+        static const char *typeName();
+
+        void set_lambda(OpenMM::Context &context, double lam_val) const;
+
+        void set_force_index(const QString &force, int index);
+
+        void add_perturbable_molecule(const OpenMMMolecule &molecule,
+                                      const QHash<QString, qint32> &start_indicies);
+
+    protected:
+        /** Map from a forcefield name to its index in the associated System */
+        QHash<QString, int> name_to_ffidx;
+    };
+}
+
+Q_DECLARE_METATYPE(SireOpenMM::LambdaLever)
+
+SIRE_END_HEADER
+
+#endif
